Serialization feature, move group to specs
