+++ specs_derive/examples/derive.rs
@@ -7,8 +7,6 @@ extern crate specs_derive;
 struct SomeGroup {
     #[group(no_serialize)]
     field1: u32,
-
-    #[group(subgroup)]
     field2: f32,
     field3: String,
 }
