+++ specs_derive/src/lib.rs
@@ -7,16 +7,20 @@ extern crate syn;
 #[macro_use]
 extern crate quote;
 extern crate specs;
+#[cfg(feature="serialize")]
 extern crate serde;
 
 use proc_macro::TokenStream;
 
-mod component;
+mod component_group;
 
 #[proc_macro_derive(ComponentGroup, attributes(group))]
-pub fn component_group(input: TokenStream) -> TokenStream {
+pub fn derive_component(input: TokenStream) -> TokenStream {
     let input = syn::parse_derive_input(&input.to_string()).unwrap();
-    match component::expand_group(&input) {
+
+    println!("Output: {:?}", component_group::expand_group(&input));
+
+    match component_group::expand_group(&input) {
         Ok(tokens) => tokens.parse().unwrap(),
         Err(err) => panic!("Error: {}", err),
     }
